factor common code in a strategy
