+++ alchemist/alchemist-cognitive-agents/src/main/kotlin/it/unibo/alchemist/model/implementations/reactions/OrientingReaction.kt
@@ -7,15 +7,17 @@
  * as described in the file LICENSE in the Alchemist distribution's top directory.
  */
 
-package it.unibo.alchemist.model.implementations.actions
+package it.unibo.alchemist.model.implementations.reactions
 
 import it.unibo.alchemist.model.cognitiveagents.characteristics.cognitive.OrientingBehavior
+import it.unibo.alchemist.model.implementations.actions.Seek2D
 import it.unibo.alchemist.model.implementations.positions.Euclidean2DPosition
-import it.unibo.alchemist.model.interfaces.Action
+import it.unibo.alchemist.model.interfaces.Environment
 import it.unibo.alchemist.model.interfaces.Node
 import it.unibo.alchemist.model.interfaces.OrientingPedestrian
 import it.unibo.alchemist.model.interfaces.Reaction
-import it.unibo.alchemist.model.interfaces.SteeringAction
+import it.unibo.alchemist.model.interfaces.Time
+import it.unibo.alchemist.model.interfaces.TimeDistribution
 import it.unibo.alchemist.model.interfaces.environments.Euclidean2DEnvironmentWithGraph
 import it.unibo.alchemist.model.interfaces.geometry.euclidean.twod.ConvexPolygon
 import it.unibo.alchemist.model.interfaces.geometry.euclidean.twod.Euclidean2DConvexShape
@@ -23,33 +25,31 @@ import it.unibo.alchemist.model.interfaces.geometry.euclidean.twod.Euclidean2DTr
 import it.unibo.alchemist.model.interfaces.graph.Euclidean2DPassage
 
 /**
- * A [SteeringAction] representing the [OrientingBehavior] of a pedestrian.
- *
- * @param T the concentration type.
- * @param N the type of landmarks of the pedestrian's cognitive map.
- * @param E the type of edges of the pedestrian's cognitive map.
- * @param M the type of nodes of the navigation graph provided by the environment.
+ * A [Reaction] representing the [OrientingBehavior] of a pedestrian.
  */
-open class Orienting<T, N : Euclidean2DConvexShape, E, M : ConvexPolygon>(
+class OrientingReaction<T, N : Euclidean2DConvexShape, E, M : ConvexPolygon>(
     private val environment: Euclidean2DEnvironmentWithGraph<*, T, M, Euclidean2DPassage>,
-    reaction: Reaction<T>,
-    pedestrian: OrientingPedestrian<T, Euclidean2DPosition, Euclidean2DTransformation, N, E>
-) : AbstractSteeringAction<T, Euclidean2DPosition>(environment, reaction, pedestrian) {
+    private val pedestrian: OrientingPedestrian<T, Euclidean2DPosition, Euclidean2DTransformation, N, E>,
+    timeDistribution: TimeDistribution<T>
+) : AbstractReaction<T>(pedestrian, timeDistribution) {
 
     private val behavior: OrientingBehavior<T, N, E, M> = OrientingBehavior(environment, pedestrian)
 
+    override fun updateInternalStatus(curTime: Time, executed: Boolean, env: Environment<T, *>) = behavior.update()
+
     @Suppress("UNCHECKED_CAST")
-    override fun cloneAction(n: Node<T>, r: Reaction<T>): Action<T> {
+    override fun cloneOnNewNode(n: Node<T>, currentTime: Time): Reaction<T> {
         require(n as? OrientingPedestrian<
             T, Euclidean2DPosition, Euclidean2DTransformation, M, Euclidean2DPassage> != null) {
             "node not compatible, required: " + pedestrian.javaClass + ", found: " + n.javaClass
         }
         n as OrientingPedestrian<T, Euclidean2DPosition, Euclidean2DTransformation, M, Euclidean2DPassage>
-        return Orienting(environment, r, n)
+        return OrientingReaction(environment, n, timeDistribution)
     }
 
-    override fun nextPosition(): Euclidean2DPosition {
-        behavior.update()
-        return Seek2D(environment, reaction, pedestrian, *behavior.desiredPosition.coordinates).nextPosition
+    override fun getRate(): Double = timeDistribution.rate
+
+    override fun execute() {
+        Seek2D(environment, this, pedestrian, *behavior.desiredPosition.coordinates).execute()
     }
 }
